python/requirements.txt
pyserial==3.1.1
slackweb==1.0.5
